[shared] move back fp_is_abnormal to CrossSectionKernels.cc (which is built with no fast math in Makefile!)
Revert "[shared] move fp_is_abnormal and related inline functions from CrossSectionKernels.cc to abnormalfp.h"
This reverts commit 3fed566e95a96035aa97d67f718eb4b0b2a0e8b0.
