--- epochX/cudacpp/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/abnormalfp.h
@@ -1,1 +1,0 @@
-../abnormalfp.h
